# webconsole-back/swagger/SGDSI_Website4sgCore_API_ADMIN_PROD.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "Website4sg Core API ADMIN PROD documentation",
    "version" : "1.3",
    "title" : "SGDSI Website4sg Core API ADMIN PROD"
  },
  "host" : "core.website4sg.saint-gobain.com",
  "basePath" : "/",
...
  "schemes" : [ "https" ],
  "paths" : {
...
    "/v1/hosting/docroots/{docrootCode}/env/{environmentCode}/servers" : {
      "get" : {
        "tags" : [ "v1hostingdocroots" ],
        "summary" : "Get all servers associated to the specified environment / docroot ",
        "description" : "By default, fetchLimit parameter is set to 25",
        "operationId" : "getAllServers",
        "parameters" : [ {
          "name" : "docrootCode",
          "in" : "path",
          "description" : "Docroot Id",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "environmentCode",
          "in" : "path",
          "description" : "Environment Id",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "search",
          "in" : "query",
          "description" : "The search engine. Available value : ssh => to get ssh servers only",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "page",
          "in" : "query",
          "description" : "Results page you want to retrieve (0..N), default value = O",
          "required" : false,
          "type" : "integer"
        }, {
          "name" : "size",
          "in" : "query",
          "description" : "Number of records per page, default value = 20",
          "required" : false,
          "type" : "integer"
        }, {
          "name" : "sort",
          "in" : "query",
          "description" : "Sorting criteria in the format: property(,asc|desc). Default sort order is ascending. Multiple sort criteria are supported.",
          "required" : false,
          "type" : "array",
          "items" : {
            "type" : "string"
          },
          "collectionFormat" : "multi"
        } ],
        "responses" : {
          "200" : {
            "description" : "Request Successfully completed",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/ServerHeader"
              }
            }
          },
          "400" : {
            "description" : "Bad Request"
          },
          "401" : {
            "description" : "Unauthorized"
          },
          "403" : {
            "description" : "Forbidden"
          },
          "500" : {
            "description" : "Internal Error"
          }
        },
        "security" : [ {
          "jwt" : [ ]
        } ]
      }
    },
    "/v1/hosting/docroots/{docrootCode}/env/{environmentCode}/servers/{hostname}" : {
      "put" : {
        "tags" : [ "v1hostingdocroots" ],
        "summary" : "Add the specified server to the specified docroot / environment",
        "description" : "",
        "operationId" : "addServer",
        "parameters" : [ {
          "name" : "docrootCode",
          "in" : "path",
          "description" : "Docroot Id",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "environmentCode",
          "in" : "path",
          "description" : "Environment Id",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "hostname",
          "in" : "path",
          "description" : "Hostname of the server to add",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "Request Successfully completed",
            "schema" : {
              "$ref" : "#/definitions/ServerDetailBody"
            }
          },
          "401" : {
            "description" : "Unauthorized"
          },
          "403" : {
            "description" : "Forbidden"
          },
          "404" : {
            "description" : "Not Found"
          },
          "500" : {
            "description" : "Internal Error"
          }
        },
        "security" : [ {
          "jwt" : [ ]
        } ]
      },
      "delete" : {
        "tags" : [ "v1hostingdocroots" ],
        "summary" : "Remove the specified server to the specified docroot / environment",
        "description" : "",
        "operationId" : "removeServer",
        "parameters" : [ {
          "name" : "docrootCode",
          "in" : "path",
          "description" : "Docroot Id",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "environmentCode",
          "in" : "path",
          "description" : "Environment Id",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "hostname",
          "in" : "path",
          "description" : "Hostname of the server to add",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "Request Successfully completed"
          },
          "401" : {
            "description" : "Unauthorized"
          },
          "403" : {
            "description" : "Forbidden"
          },
          "404" : {
            "description" : "Not Found"
          },
          "500" : {
            "description" : "Internal Error"
          }
        },
        "security" : [ {
          "jwt" : [ ]
        } ]
      }
    },
...
    "/v1/servers" : {
      "get" : {
        "tags" : [ "v1servers" ],
        "summary" : "Get all servers",
        "description" : "",
        "operationId" : "getAllServers",
        "parameters" : [ {
          "name" : "name",
          "in" : "query",
          "description" : "The name of environment",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "search",
          "in" : "query",
          "description" : "The search engine",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "page",
          "in" : "query",
          "description" : "Results page you want to retrieve (0..N), default value = O",
          "required" : false,
          "type" : "integer"
        }, {
          "name" : "size",
          "in" : "query",
          "description" : "Number of records per page, default value = 20",
          "required" : false,
          "type" : "integer"
        }, {
          "name" : "sort",
          "in" : "query",
          "description" : "Sorting criteria in the format: property(,asc|desc). Default sort order is ascending. Multiple sort criteria are supported.",
          "required" : false,
          "type" : "array",
          "items" : {
            "type" : "string"
          },
          "collectionFormat" : "multi"
        } ],
        "responses" : {
          "200" : {
            "description" : "Request Successfully completed",
            "schema" : {
              "$ref" : "#/definitions/PageServerDetailBody"
            }
          },
          "401" : {
            "description" : "Unauthorized"
          },
          "403" : {
            "description" : "Forbidden"
          },
          "404" : {
            "description" : "Not Found"
          },
          "500" : {
            "description" : "Internal Error"
          }
        }
      }
    },
    "/v1/servers/{hostname}" : {
      "get" : {
        "tags" : [ "v1servers" ],
        "summary" : "Get server by hostname",
        "description" : "",
        "operationId" : "getServer",
        "parameters" : [ {
          "name" : "hostname",
          "in" : "path",
          "description" : "host name",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "Request Successfully completed",
            "schema" : {
              "$ref" : "#/definitions/ServerDetailBody"
            }
          },
          "401" : {
            "description" : "Unauthorized"
          },
          "403" : {
            "description" : "Forbidden"
          },
          "404" : {
            "description" : "Not Found"
          },
          "500" : {
            "description" : "Internal Error"
          }
        }
      },
      "put" : {
        "tags" : [ "v1servers" ],
        "summary" : "Update or add a server",
        "description" : "",
        "operationId" : "createOrUpdateServer",
        "parameters" : [ {
          "name" : "hostname",
          "in" : "path",
          "description" : "host name",
          "required" : true,
          "type" : "string"
        }, {
          "in" : "body",
          "name" : "body",
          "description" : "Server object",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/Server"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Request Successfully completed",
            "schema" : {
              "$ref" : "#/definitions/ServerDetailBody"
            }
          },
          "401" : {
            "description" : "Unauthorized"
          },
          "403" : {
            "description" : "Forbidden"
          },
          "404" : {
            "description" : "Not Found"
          },
          "500" : {
            "description" : "Internal Error"
          }
        }
      }
    },
...
  },
  "definitions" : {
...
    "PageServerDetailBody" : {
      "type" : "object",
      "properties" : {
        "totalElements" : {
          "type" : "integer",
          "format" : "int64"
        },
        "totalPages" : {
          "type" : "integer",
          "format" : "int32"
        },
        "sort" : {
          "$ref" : "#/definitions/Sort"
        },
        "size" : {
          "type" : "integer",
          "format" : "int32"
        },
        "content" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/ServerDetailBody"
          }
        },
        "number" : {
          "type" : "integer",
          "format" : "int32"
        },
        "first" : {
          "type" : "boolean"
        },
        "numberOfElements" : {
          "type" : "integer",
          "format" : "int32"
        },
        "last" : {
          "type" : "boolean"
        },
        "pageable" : {
          "$ref" : "#/definitions/Pageable"
        },
        "empty" : {
          "type" : "boolean"
        }
      }
    },
...
    "Pageable" : {
      "type" : "object",
      "properties" : {
        "sort" : {
          "$ref" : "#/definitions/Sort"
        },
        "offset" : {
          "type" : "integer",
          "format" : "int64"
        },
        "paged" : {
          "type" : "boolean"
        },
        "unpaged" : {
          "type" : "boolean"
        },
        "pageNumber" : {
          "type" : "integer",
          "format" : "int32"
        },
        "pageSize" : {
          "type" : "integer",
          "format" : "int32"
        }
      }
    },
...
    "Server" : {
      "type" : "object",
      "required" : [ "domain" ],
      "properties" : {
        "domain" : {
          "type" : "string"
        },
        "enable" : {
          "type" : "boolean"
        },
        "sshServer" : {
          "type" : "boolean"
        },
        "login" : {
          "type" : "string"
        },
        "hostname" : {
          "type" : "string"
        },
        "created" : {
          "type" : "string",
          "format" : "date-time"
        },
        "lastUpdate" : {
          "type" : "string",
          "format" : "date-time"
        }
      },
      "description" : "Server Object"
    },
    "ServerDetailBody" : {
      "type" : "object",
      "required" : [ "domain", "hostname" ],
      "properties" : {
        "domain" : {
          "type" : "string"
        },
        "enable" : {
          "type" : "boolean"
        },
        "sshServer" : {
          "type" : "boolean"
        },
        "login" : {
          "type" : "string"
        },
        "hostname" : {
          "type" : "string"
        },
        "created" : {
          "type" : "string",
          "format" : "date-time"
        },
        "lastUpdate" : {
          "type" : "string",
          "format" : "date-time"
        }
      },
      "description" : "ServerDetailBody Object"
    },
    "ServerHeader" : {
      "type" : "object",
      "required" : [ "hostname" ],
      "properties" : {
        "hostname" : {
          "type" : "string"
        }
      },
      "description" : "ServerHeader Object"
    },
...
    "Sort" : {
      "type" : "object",
      "properties" : {
        "sorted" : {
          "type" : "boolean"
        },
        "unsorted" : {
          "type" : "boolean"
        },
        "empty" : {
          "type" : "boolean"
        }
      }
    },
...
  }
}